morningfocus-web/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "morningfocus-web"
version = "0.1.0"
description = "MorningFocus — web UI with SQLite"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "sqlalchemy>=2.0",
    "python-dateutil>=2.8",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = ["httpx>=0.27", "pytest>=8", "pytest-asyncio>=0.23"]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
